Refactored add allocation form to by popup when clicking 'New' to clean up layout
