+++ budget-app-react/src/components/AddAllocation.jsx
@@ -1,6 +1,7 @@
 import { useState } from "react";
+import CancelButton from "./CancelButton";
 
-export default function AddIncome({ fetchAllocationData }) {
+export default function AddAllocation({ fetchAllocationData, setAddMode }) {
 
     const [allocationDescription, setAllocationDescription] = useState("");
     const [allocationAmount, setAllocationAmount] = useState("");
@@ -35,32 +36,37 @@ export default function AddIncome({ fetchAllocationData }) {
             .then(setAllocationDescription(""))
             .then(setAllocationAmount(""))
             .then(fetchAllocationData)
+            .then(setAddMode(false))
 
     }
 
     return (
-        <form onSubmit={handleSubmit}>
-            <label htmlFor="allocation-description">Allocation</label>
-            <input
-                type="text"
-                name="allocation-description"
-                value={allocationDescription}
-                onChange={(e) => setAllocationDescription(e.target.value)}
-                placeholder="Add a description..."
-            />
+        <div className="edit-form">
+            <form onSubmit={handleSubmit} className="edit-form-inner inner-allocation">
+                <label htmlFor="allocation-description">Allocation</label>
+                <input
+                    type="text"
+                    name="allocation-description"
+                    value={allocationDescription}
+                    onChange={(e) => setAllocationDescription(e.target.value)}
+                    placeholder="Add a description..."
+                />
 
-            <label htmlFor="allocation-amount">Amount</label>
-            <input
-                type="number"
-                name="allocation-amount"
-                value={allocationAmount}
-                onChange={(e) => setAllocationAmount(e.target.value)}
-            />
+                <label htmlFor="allocation-amount">Amount</label>
+                <input
+                    type="number"
+                    name="allocation-amount"
+                    value={allocationAmount}
+                    onChange={(e) => setAllocationAmount(e.target.value)}
+                />
 
-            <button className="add-button">Add</button>
+                <br />
+                <button className="add-button">Add</button>
+                <CancelButton setMode={setAddMode} />
 
-            {descriptionError && <div>Allocation description cannot be blank</div>}
-            {ammountError && <div>Amount must be least $1.00</div>}
-        </form>
+                {descriptionError && <div>Allocation description cannot be blank</div>}
+                {ammountError && <div>Amount must be least $1.00</div>}
+            </form>
+        </div>
     );
 }
