+++ budget-app-react/src/components/AddButton.jsx
@@ -1,0 +1,7 @@
+
+export default function AddButton({ setAddMode }) {
+
+    return (
+        <button onClick={() => setAddMode(true)}>New</button>
+    )
+}
